requirements.txt
google-genai
python-dotenv
flask
flask-cors
pdfplumber
python-docx
pandas
Pillow
docling
openpyxl
